master: Add --preseed option
Sometimes it is necessary to make a release of a package against some
old development packages. Since those are not available in the general
repository, the script now provides a --preseed option which will
additionally look in the folder preseed for additional packages. When
resolving dependencies, those will take precedence over the packages
found in the other repositories, except for pbuilder-result